backend/requirements.txt
openai>=1.0.0
groq
web3
httpx
python-dotenv
geopandas
shapely
fiona
pyproj
rasterio
fastapi
uvicorn[standard]
pydantic
requests
python-multipart
psycopg2-binary
pgvector
sqlalchemy
numpy